sovle problem 3 and do some refactoring

+++ IProblem.cs
@@ -2,6 +2,6 @@
 {
     interface IProblem
     {
-        int Run();
+        long Run();
     }
 }
