+++ Problems/Problem2.cs
@@ -5,7 +5,7 @@ namespace ProjectEuler.Problems
 {
     class Problem2 : IProblem
     {
-        public int Run()
+        public long Run()
         {
             return new Fibonacci(4000000).Where(n => n.IsDivisiableBy(2)).Sum();
         }
